+++ src/components/manage/Form/InlineForm.jsx
@@ -75,7 +75,11 @@ const InlineForm = ({
             id={field}
             fieldSet={defaultFieldset.title.toLowerCase()}
             focus={index === 0}
-            value={formData[field] || schema.properties[field].default}
+            value={
+              'default' in schema.properties[field]
+                ? formData[field] || schema.properties[field].default
+                : formData[field]
+            }
             required={schema.required.indexOf(field) !== -1}
             onChange={(id, value) => {
               onChangeField(id, value);
@@ -97,7 +101,11 @@ const InlineForm = ({
               <Field
                 {...schema.properties[field]}
                 id={field}
-                value={formData[field] || schema.properties[field].default}
+                value={
+                  'default' in schema.properties[field]
+                    ? formData[field] || schema.properties[field].default
+                    : formData[field]
+                }
                 required={schema.required.indexOf(field) !== -1}
                 onChange={(id, value) => {
                   onChangeField(id, value);